Replace analytics reporting devices with `Firebase.DatabaseAPI`
