+++ Sources/CLI/CommandLineTool.swift
@@ -1,6 +1,7 @@
 import Foundation
 import SwiftCLI
 import Configuration
+import Firebase
 import Git
 
 public final class CommandLineTool {
@@ -81,8 +82,8 @@ public final class CommandLineTool {
 		}
 
 		Logger.step("Reporting error")
-		let reporter = ErrorAnalyticsReporter(firebaseProjectId: firebaseProjectId, firebaseSecretToken: firebaseSecretToken)
-		try reporter.report(error.analyticsData(startDate: startDate))
+		let reporter = Firebase.DatabaseAPI(firebaseProjectId: firebaseProjectId, firebaseSecretToken: firebaseSecretToken)
+		try reporter.post(documentName: "errors", body: error.analyticsData(startDate: startDate))
 	}
 
 	private func logElapsedTime(withStartDate startDate: Date) {
